backend/task_management/requirements.txt
# Task Management AI System Requirements

# Core Dependencies
flask>=2.3.0
flask-cors>=4.0.0
openai>=1.0.0
pymysql>=1.1.0
python-dotenv>=1.0.0

# Database and Vector Storage
chromadb>=0.4.0
faiss-cpu>=1.7.4

# NLP Processing
spacy>=3.6.0
numpy>=1.24.0

# Scheduling
apscheduler>=3.10.0

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Logging and Monitoring
structlog>=23.0.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
requests>=2.31.0

# Optional: GPU support for FAISS (uncomment if needed)
# faiss-gpu>=1.7.4

# Optional: Advanced NLP models (uncomment if needed)
# transformers>=4.30.0
# torch>=2.0.0